feat: add animations to all sections, improve mobile responsiveness, performance, and accessibility
