+++ src/components/Education.tsx
@@ -1,9 +1,13 @@
+"use client";
+
 import { useContext } from "react";
-import { CiSaveDown1 } from "react-icons/ci";
 import { LayoutContext } from "./context";
+import { CiSaveDown1 } from "react-icons/ci";
+import { motion } from "framer-motion";
 
 export default function Timeline() {
   const context = useContext(LayoutContext);
+
   if (!context) {
     throw new Error(
       "LayoutContext must be used within a LayoutContext.Provider"
@@ -13,21 +17,34 @@ export default function Timeline() {
   const { translations, isRTL } = context;
   const educations = translations?.education || {};
 
-  // Sort years in descending order
   const sortedEducations = Object.entries(educations).sort((a, b) => {
     return Number(b[0]) - Number(a[0]);
   });
 
   return (
-    <div className="w-full py-12">
+    <section className="w-full py-12">
       <h2 className="text-start mb-12 font-bold text-4xl">
         {translations.Educationheading}
       </h2>
 
-      <div className="relative border-t-2 border-black dark:border-white flex flex-wrap justify-between items-start gap-y-12">
-        {sortedEducations.map(([year, edu]) => (
-          <div
-            key={year}
+      <motion.div
+        className="relative border-t-2 border-black dark:border-white flex flex-wrap justify-between items-start gap-y-12"
+        initial="hidden"
+        whileInView="visible"
+        viewport={{ once: true, amount: 0.3 }}
+        transition={{ staggerChildren: 0.2 }}
+      >
+        {sortedEducations.map(([year, edu], index) => (
+          <motion.div
+            key={index}
+            variants={{
+              hidden: { opacity: 0, y: 60 },
+              visible: {
+                opacity: 1,
+                y: 0,
+                transition: { duration: 0.6, ease: "easeOut" },
+              },
+            }}
             className={`relative text-center w-full md:w-1/4 px-4 transition duration-300 ${
               isRTL ? "text-right" : ""
             }`}
@@ -38,7 +55,7 @@ export default function Timeline() {
             </div>
 
             {/* Card */}
-            <div className="mt-6 border border-gray-200 dark:border-gray-700 rounded p-4 text-left">
+            <div className="mt-6 border border-gray-200 dark:border-gray-700 rounded p-4 text-left min-h-[220px] bg-white dark:bg-gray-900 shadow-sm hover:shadow-md transition duration-300">
               <h3 className="font-bold text-sm text-gray-500">{year}</h3>
               <p className="font-semibold text-lg text-black dark:text-white mt-1">
                 {edu.title}
@@ -59,9 +76,9 @@ export default function Timeline() {
                 </p>
               )}
             </div>
-          </div>
+          </motion.div>
         ))}
-      </div>
-    </div>
+      </motion.div>
+    </section>
   );
 }
